requirements/requirements.txt
apache-superset==2.1.3
pandas
numpy
matplotlib
seaborn
scikit-learn
psycopg2-binary
